# docs/swagger.yaml
basePath: /api/albums/feed
definitions:
  http.Error:
    properties:
      message:
        type: string
    type: object
...
  http.defaultResponse:
    properties:
      status:
        type: string
    type: object
...
  http.playlistCreateInput:
    properties:
      description:
        type: string
      name:
        type: string
      users:
        items:
          type: integer
        type: array
    type: object
  http.playlistCreateResponse:
    properties:
      id:
        type: integer
    type: object
  http.playlistUpdateInput:
    properties:
      description:
        type: string
      name:
        type: string
      users:
        items:
          type: integer
        type: array
    type: object
...
  models.Date:
    properties:
      time.Time:
        type: string
    type: object
  models.PlaylistTransfer:
    properties:
      cover:
        type: string
      description:
        type: string
      id:
        type: integer
      isLiked:
        type: boolean
      name:
        type: string
      users:
        items:
          $ref: '#/definitions/models.UserTransfer'
        type: array
    type: object
  models.Sex:
    enum:
    - M
    - F
    - O
    type: string
    x-enum-varnames:
    - Male
    - Female
    - Other
...
  models.UserTransfer:
    properties:
      avatarSrc:
        type: string
      birthDate:
        $ref: '#/definitions/models.Date'
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      sex:
        $ref: '#/definitions/models.Sex'
      username:
        type: string
    type: object
host: localhost:4443
info:
  contact:
    email: yarik1448kuzmin@gmail.com
    name: Fluire API Support
  description: Server API for Fluire Streaming Service Application
  title: Fluire API
  version: 1.0.1
paths:
...
  /api/playlists/:
    post:
      consumes:
      - application/json
      description: Create new playlist by sent object
      parameters:
      - description: Playlist info
        in: body
        name: playlist
        required: true
        schema:
          $ref: '#/definitions/http.playlistCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: Playlist created
          schema:
            $ref: '#/definitions/http.playlistCreateResponse'
        "400":
          description: Incorrect input
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: User hasn't rights
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Create Playlist
      tags:
      - Playlist
  /api/playlists/{playlistID}/:
    delete:
      description: Delete playlist with chosen ID
      produces:
      - application/json
      responses:
        "200":
          description: Playlist deleted
          schema:
            $ref: '#/definitions/http.defaultResponse'
        "400":
          description: Client error
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: User hasn't rights
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Delete Playlist
      tags:
      - Playlist
    get:
      description: Get playlist with chosen ID
      produces:
      - application/json
      responses:
        "200":
          description: Playlist got
          schema:
            $ref: '#/definitions/models.PlaylistTransfer'
        "400":
          description: Incorrect input
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Get Playlist
      tags:
      - Playlist
  /api/playlists/{playlistID}/cover:
    post:
      consumes:
      - multipart/form-data
      description: Update playlist cover
      parameters:
      - description: Cover file
        in: formData
        name: cover
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Cover updated
          schema:
            $ref: '#/definitions/http.defaultResponse'
        "400":
          description: Invalid form data
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User Unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: User hasn't rights
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Upload Cover
      tags:
      - Playlist
  /api/playlists/{playlistID}/like:
    post:
      description: Set like by user to chosen playlist (add to favorite)
      produces:
      - application/json
      responses:
        "200":
          description: Like set
          schema:
            $ref: '#/definitions/http.defaultResponse'
        "400":
          description: Client error
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Set like
      tags:
      - Playlist
...
  /api/playlists/{playlistID}/tracks/{trackID}:
    delete:
      description: Delete track from playlist
      produces:
      - application/json
      responses:
        "200":
          description: Track deleted
          schema:
            $ref: '#/definitions/http.playlistCreateResponse'
        "400":
          description: Incorrect input
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: User hasn't rights
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Delete Track
      tags:
      - Playlist
    post:
      description: Add track into playlist
      produces:
      - application/json
      responses:
        "200":
          description: Track added
          schema:
            $ref: '#/definitions/http.playlistCreateResponse'
        "400":
          description: Incorrect input
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: User hasn't rights
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Add Track
      tags:
      - Playlist
  /api/playlists/{playlistID}/unlike:
    post:
      description: Remove like by user from chosen playlist (remove from favorite)
      produces:
      - application/json
      responses:
        "200":
          description: Like removed
          schema:
            $ref: '#/definitions/http.defaultResponse'
        "400":
          description: Client error
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Remove like
      tags:
      - Playlist
  /api/playlists/{playlistID}/update:
    post:
      consumes:
      - application/json
      description: Update playlist
      parameters:
      - description: Playlist info
        in: body
        name: playlist
        required: true
        schema:
          $ref: '#/definitions/http.playlistUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: Playlist updated
          schema:
            $ref: '#/definitions/http.defaultResponse'
        "400":
          description: Client error
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: User hasn't rights
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Update Playlist
      tags:
      - Playlist
  /api/playlists/feed:
    get:
      description: Feed playlists
      produces:
      - application/json
      responses:
        "200":
          description: Playlist feed
          schema:
            items:
              $ref: '#/definitions/models.PlaylistTransfer'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Playlist Feed
      tags:
      - Feed
...
  /api/users/{userID}/favorite/playlists:
    get:
      description: Get user's favorite playlists
      produces:
      - application/json
      responses:
        "200":
          description: Playlists got
          schema:
            items:
              $ref: '#/definitions/models.PlaylistTransfer'
            type: array
        "400":
          description: Incorrect input
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: Forbidden user
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Favorite Playlists
      tags:
      - Favorite
...
  /api/users/{userID}/playlists:
    get:
      description: All playlists of user with chosen ID
      produces:
      - application/json
      responses:
        "200":
          description: Show playlists
          schema:
            items:
              $ref: '#/definitions/models.PlaylistTransfer'
            type: array
        "400":
          description: Client error
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Playlists of User
      tags:
      - User
...
swagger: "2.0"